Move all task related tests to a single file and add taskResult tests
